feat: refactored profile page
