+++ apps/frontend/src/app/profile/[id]/page.tsx
@@ -1,4 +1,4 @@
-import ProfilePageComponent from '@/components/profile-page';
+import ProfilePageComponent from '@/components/profile/profile-page';
 
 export default function ProfilePage() {
   return <ProfilePageComponent />;
